# Conflicts:
#	Project/Assets/LifeDeadMain/Scripts/DataModule/DataController.cs
#	Project/Assets/StreamingAssets/DataJson.json
#	Project/Library/CurrentLayout.dwlt
#	Project/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll.mdb
#	Project/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	Project/Library/assetDatabase3
#	Project/Library/expandedItems
#	Project/Library/metadata/00/00000000000000004000000000000000
#	Project/Library/metadata/fd/fdab7a9af34477d46bece06f63f7720a
#	Project/Temp/__Backupscenes/0.backup

+++ Project/Assets/LifeDeadMain/Scripts/DataModule/LevelData.cs
@@ -12,7 +12,7 @@ using System.Collections;
 public class LevelData
 {
     public int Name { set; get; }
-    public float Time { set; get; }
+    public int Time { set; get; }
     public int Score { set; get; }
     public int StarNum { set; get; }    	
 }
